fix(chaincode): Rename *_ops to *-ops
This patch renames chaincode_ops and channel_ops to chaincode-ops and
channel-ops. The reason for this modification is that k8s chaincode builder
does not support chaincode names containing underscore (This is due to
k8s's metadata only supports DNS-1123 names).

This patch also includes minor fixes on chaincodes.
