+++ chaincode/.gitignore
@@ -1,4 +1,4 @@
-chaincode_ops/chaincode_ops
-channel_ops/channel_ops
+chaincode-ops/chaincode-ops
+channel-ops/channel-ops
 /**/coverage.html
 /**/coverprofile.out
