+++ chaincode/chaincode-ops/core/mocks/chaincodestub.go
