+++ chaincode/chaincode-ops/go.mod
@@ -1,4 +1,4 @@
-module github.com/hyperledger-labs/fabric-opssc/chaincode/chaincode_ops
+module github.com/hyperledger-labs/fabric-opssc/chaincode/chaincode-ops
 
 go 1.14
