+++ chaincode/channel-ops/chaincode/mocks/chaincodestub.go
