+++ chaincode/channel-ops/chaincode/mocks/statequeryiterator.go
