+++ opssc-agent/README.md
@@ -50,8 +50,8 @@ The following environment variables must be set:
 | Hyperledger Fabric | `DISCOVER_AS_LOCALHOST` | `false`                                      | Whether to discover as localhost                                                                           |
 | Hyperledger Fabric | `CONNECTION_PROFILE`    | `/opt/fabric/config/connection-profile.yaml` | Connection profile path for the organization (NOTE: should be written all peers owned by the organization) |
 | OpsSC              | `CHANNEL_NAME`          | `ops-channel`                                | Channel name for the OpsSC                                                                                 |
-| OpsSC              | `CC_OPS_CC_NAME`        | `chaincode_ops`                              | Chaincode name of the chaincode OpsSC                                                                      |
-| OpsSC              | `CH_OPS_CC_NAME`        | `channel_ops`                                | Chaincode name of the channel OpsSC                                                                        |
+| OpsSC              | `CC_OPS_CC_NAME`        | `chaincode-ops`                              | Chaincode name of the chaincode OpsSC                                                                      |
+| OpsSC              | `CH_OPS_CC_NAME`        | `channel-ops`                                | Chaincode name of the channel OpsSC                                                                        |
 | Chaincode Ops      | `GIT_USER`              | None                                         | Git user to access to the chaincode repository (If not set, access without credentials)                    |
 | Chaincode Ops      | `GIT_PASSWORD`          | None                                         | Git password to access to the chaincode repository (If not set, access without credentials)                |
 | Chaincode Ops      | `GOPATH`                | None                                         | GOPATH                                                                                                     |
