+++ opssc-api-server/charts/opssc-api-server/values.yaml
@@ -69,8 +69,8 @@ api:
   utilEnabled: "true"
 
 opsscChannelName: "ops-channel"
-opsscChannelOpsCCName: "channel_ops"
-opsscChaincodeOpsCCName: "chaincode_ops"
+opsscChannelOpsCCName: "channel-ops"
+opsscChaincodeOpsCCName: "chaincode-ops"
 
 secrets:
   connectionProfileItemKeyName: HLF_CONNECTION_PROFILE_ORG1
